# backend/swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/books/languages/static/{alpha2}": {
      "get": {
        "operationId": "LanguagesController_getOneStatic",
        "summary": "Get a language by its alpha2 code",
        "description": "Get a language by its alpha2 code.",
        "parameters": [
          {
            "name": "alpha2",
            "required": true,
            "in": "path",
            "example": "ES",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The language.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ES": {
                      "type": "string",
                      "example": "spanish"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Books/Languages"
        ]
      }
    },
    "/api/books/languages/{id}": {
      "get": {
        "operationId": "LanguagesController_findOne",
        "summary": "Get a language by id",
        "description": "Get a language by id.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The language.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Language"
                }
              }
            }
          }
        },
        "tags": [
          "Books/Languages"
        ]
      },
      "delete": {
        "operationId": "LanguagesController_delete",
        "summary": "Delete a language by id",
        "description": "Delete a language by id.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "The language has been deleted."
          }
        },
        "tags": [
          "Books/Languages"
        ]
      }
    },
...
    "/api/books/locations": {
      "post": {
        "operationId": "LocationsController_create",
        "summary": "Create a location",
        "description": "[Admin, Store] Create a location associated to the store.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLocationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The location.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Location"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized."
          },
          "403": {
            "description": "Forbidden."
          },
          "404": {
            "description": "Store not found."
          }
        },
        "tags": [
          "Books/Locations"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "LocationsController_findMany",
        "summary": "Get all the locations",
        "description": "[Admin, Store] Get all the locations associated to the store.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "The locations.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Location"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized."
          },
          "403": {
            "description": "Forbidden."
          },
          "404": {
            "description": "Store not found."
          }
        },
        "tags": [
          "Books/Locations"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/books/locations/{id}": {
      "get": {
        "operationId": "LocationsController_findOne",
        "summary": "Get a location by id",
        "description": "[Admin, Store] Get a location by id.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The location.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Location"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized."
          },
          "403": {
            "description": "Forbidden."
          },
          "404": {
            "description": "Store not found."
          }
        },
        "tags": [
          "Books/Locations"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "LocationsController_update",
        "summary": "Update a location",
        "description": "[Admin, Store] Update a location associated to the store.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateLocationDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "The location has been updated."
          },
          "401": {
            "description": "Unauthorized."
          },
          "403": {
            "description": "Forbidden."
          },
          "404": {
            "description": "Store not found."
          },
          "500": {
            "description": "Location not updated."
          }
        },
        "tags": [
          "Books/Locations"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "LocationsController_delete",
        "summary": "Delete a location",
        "description": "[Admin, Store] (Hard) Delete a location associated to the store.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "The location has been (hard) deleted."
          },
          "401": {
            "description": "Unauthorized."
          },
          "403": {
            "description": "Forbidden."
          },
          "404": {
            "description": "Store not found."
          },
          "500": {
            "description": "Location not deleted."
          }
        },
        "tags": [
          "Books/Locations"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Librería Santa Bárbara API",
    "description": "API documentation for Librería Santa Bárbara",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "Language": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1,
            "description": "The id of the language"
          },
          "book_id": {
            "type": "number",
            "example": 1,
            "description": "The id of the book that the language references to"
          },
          "language": {
            "type": "string",
            "example": "EN",
            "description": "ISO 639-1 Alpha-2 code of the language"
          }
        },
        "required": [
          "id",
          "book_id",
          "language"
        ]
      },
...
      "CreateLocationDto": {
        "type": "object",
        "properties": {
          "location": {
            "type": "string",
            "example": "Storage 1",
            "description": "The location of the book"
          }
        },
        "required": [
          "location"
        ]
      },
      "Location": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1,
            "description": "The id of the location"
          },
          "store_id": {
            "type": "number",
            "example": 1,
            "description": "The id of the store where the location is"
          },
          "location": {
            "type": "string",
            "example": "Storage",
            "description": "The location"
          }
        },
        "required": [
          "id",
          "store_id",
          "location"
        ]
      },
      "UpdateLocationDto": {
        "type": "object",
        "properties": {
          "location": {
            "type": "string",
            "example": "Storage 1",
            "description": "The location of the book"
          }
        }
      },
...
    }
  }
}